diffusion/Latent Diffusion/Conditioned/CelebA/celeba-cond-ddpm-model/requirements.txt
torch
torchvision
transformers
gradio>=3.16
spacy
datasets
Pillow